requirements.txt
appdirs==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bitstring==3.1.7
bleach==3.3.0
boto3==1.17.49
botocore==1.20.49
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
datapackage==1.15.2
decorator==4.4.2
defusedxml==0.7.1
dgl==0.6.1
dgl-cu102==0.7a210410
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
et-xmlfile==1.0.1
greenlet==1.0.0
html5lib==1.0.1
idna==2.8
ijson==3.1.4
ipaddr==2.2.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonlines==2.0.0
jsonpointer==2.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kvfile==0.0.12
linear-tsv==1.1.0
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
msgpack==0.6.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.20.2
openpyxl==3.0.7
packaging==20.3
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
progress==1.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.22.0
retrying==1.3.3
rfc3986==1.4.0
s3transfer==0.3.6
scipy==1.6.2
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.4.7
tableschema==1.20.2
tabulate==0.8.9
tabulator==1.53.5
terminado==0.9.4
testpath==0.4.4
torch==1.8.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
xmljson==0.2.1
